Revert "feature/session config object"
