--- src/application/intents.py
@@ -1,11 +1,0 @@
-from dataclasses import dataclass
-
-
-class GameUpdate:
-    pass
-
-
-@dataclass(frozen=True)
-class CursorMove(GameUpdate):
-    dx: int
-    dy: int
